requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
pandas
scikit-learn
google-generativeai
fastapi
uvicorn
python-multipart
